Restructured ChessServer\Command
